requirements.txt
accelerate
datasets
diffusers==0.26.0
numpy==1.26.1
packaging==23.2
pandas_stubs==1.2.0.57
Pillow==10.1.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.66.1
transformers
gradio
jmespath
opencv-python
PyWavelet
gradio
huggingface_hub==0.25.2